Refactors plugin registry for internal use
Moves aggregate plugin registry to internal package.

This change prevents direct external usage of the aggregate registry and promotes better encapsulation. It cleans up package structure.
