+++ pkg/plugin/registry/internal/aggregate.go
@@ -1,4 +1,4 @@
-package registry
+package internal
 
 import (
 	"context"
