+++ pkg/plugin/registry/list.go
@@ -5,34 +5,12 @@ import (
 
 	"github.com/unmango/go/option"
 	"github.com/unstoppablemango/ux/pkg/plugin"
+	"github.com/unstoppablemango/ux/pkg/plugin/registry/list"
 )
 
-type ListOptions struct {
-	disableDefault bool
-	registries     []plugin.Registry
-}
-
-type ListOption func(*ListOptions)
-
-func List(ctx context.Context, options ...ListOption) (plugin.List, error) {
-	opts := ListOptions{}
+func List(ctx context.Context, options ...list.Option) (plugin.List, error) {
+	opts := list.Options{}
 	option.ApplyAll(&opts, options)
 
-	return opts.aggregate().List(ctx)
-}
-
-func (o ListOptions) aggregate() plugin.Registry {
-	if o.disableDefault {
-		return Aggregate(o.registries)
-	} else {
-		return allOrDefault(o.registries, nil)
-	}
-}
-
-func allOrDefault(registries []plugin.Registry, def plugin.Registry) plugin.Registry {
-	if len(registries) > 0 {
-		return Aggregate(registries)
-	} else {
-		return def
-	}
+	return opts.Aggregate().List(ctx)
 }
