+++ pkg/plugin/registry/local/local.go
@@ -6,11 +6,12 @@ import (
 	"github.com/unstoppablemango/ux/pkg/config"
 	"github.com/unstoppablemango/ux/pkg/plugin"
 	"github.com/unstoppablemango/ux/pkg/plugin/registry"
+	"github.com/unstoppablemango/ux/pkg/plugin/registry/internal"
 )
 
 var (
 	BinPattern = regexp.MustCompile(`(.+2.+)|(ux-.+)`)
 	UserConfig = registry.IgnoreNotFound(LocalDirectory(config.PluginDir))
 
-	Default plugin.Registry = registry.Aggregate{UserConfig}
+	Default plugin.Registry = internal.Aggregate{UserConfig}
 )
